DB update

+++ IspDataDb/dbo/Tables/Branch.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE [dbo].[Branch] (
-    [IDNo]          SMALLINT      IDENTITY (1, 1) NOT NULL,
+    [IDNo]          TINYINT       IDENTITY (1, 1) NOT NULL,
     [BranchCode]    VARCHAR (5)   COLLATE SQL_Latin1_General_CP1_CI_AS NOT NULL,
     [BranchName]    VARCHAR (50)  COLLATE SQL_Latin1_General_CP1_CI_AS NOT NULL,
     [BranchNameAra] NVARCHAR (50) COLLATE SQL_Latin1_General_CP1_CI_AS NOT NULL,
@@ -12,6 +12,8 @@
     CONSTRAINT [IX_BranchName] UNIQUE NONCLUSTERED ([BranchName] ASC),
     CONSTRAINT [IX_BranchNameAra] UNIQUE NONCLUSTERED ([BranchNameAra] ASC)
 );
+
+
